Rename file abstract_classes.py -> protocols.py
