+++ cg/services/run_devices/pacbio/data_transfer_service/data_transfer_service.py
@@ -4,7 +4,6 @@ import logging
 
 from pydantic import ValidationError
 
-from cg.services.run_devices.abstract_classes import PostProcessingDataTransferService
 from cg.services.run_devices.error_handler import handle_post_processing_errors
 from cg.services.run_devices.exc import (
     PostProcessingDataTransferError,
@@ -26,6 +25,7 @@ from cg.services.run_devices.pacbio.data_transfer_service.utils import (
 from cg.services.run_devices.pacbio.metrics_parser.metrics_parser import PacBioMetricsParser
 from cg.services.run_devices.pacbio.metrics_parser.models import PacBioMetrics
 from cg.services.run_devices.pacbio.run_data_generator.run_data import PacBioRunData
+from cg.services.run_devices.protocols import PostProcessingDataTransferService
 
 LOG = logging.getLogger(__name__)
